Fix featured mod deployment  (#627)
* Fix casting int to short
* Fix deployment text too long

+++ src/main/java/com/faforever/api/featuredmods/FeaturedModService.java
@@ -40,7 +40,7 @@ public class FeaturedModService {
     legacyFeaturedModFileRepository.save(modName, version, featuredModFiles);
   }
 
-  public Map<String, Short> getFileIds(String modName) {
+  public Map<String, Integer> getFileIds(String modName) {
     return legacyFeaturedModFileRepository.getFileIds(modName);
   }
